refactor: move sameOrigin into origin.ts for reuse
